[BUGFIX] Corrected Preesm issue#3: The inserted SDFForkVertex was the target of an edge with the wrong consumption rate, provoking non-schedulable graph errors.
The problem occurs when an edge stands between two vertices with the same number of repetitions and the target one is instantiated during flattening due to inner loops.
In such case a SDFForkVertex is inserted between the two vertices, but the edge between the source and the Fork is set with a consumption rate without considering the number of repetitions of the source.